ragpipeline/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.24.0
pydantic==2.10.0
google-generativeai==0.3.2
python-dotenv==1.0.0
chromadb==1.0.7
requests==2.31.0